+++ FakeNFT/Scenes /Statistics/UserCard/UserCardViewProtocol.swift
@@ -12,12 +12,3 @@ protocol UserCardViewProtocol: AnyObject {
     func showView(viewController: UIViewController)
     func setupUserData(with user: User)
 }
-
-protocol UserCardPresenterProtocol {
-    var view: UserCardViewProtocol? { get set }
-    func viewDidLoad()
-    func setUser(with user: User)
-    func countUserNFTS() -> Int
-    func userWebsiteButtonTapped()
-    func presentNftCollectionTapped()
-}
