+++ FakeNFT/Scenes /Statistics/UsersCollection/UsersCollectionPresenter.swift
@@ -26,76 +26,6 @@ class UsersCollectionPresenter {
         self.serviceCart = serviceCart
         self.serviceProfile = serviceProfile
     }
-}
-
-extension UsersCollectionPresenter: UsersCollectionPresenterProtocol {
-    func viewDidLoad() {
-        loadNfts()
-        updateCart()
-        loadProfile()
-    }
-
-    func setNftsId(nftsId: [String]) {
-        self.nftsId = nftsId
-    }
-
-    func nftsCount() -> Int {
-        return nfts.count
-    }
-
-    func nftForIndex(indexPath: IndexPath) -> Nft {
-        return nfts[indexPath.row]
-    }
-
-    func cartContainsNft(nft: Nft) -> Bool {
-        nftsInCart.contains(where: { $0.id == nft.id })
-    }
-
-    func isInFavourites(nft: Nft) -> Bool {
-        nftsInFavourites.contains(where: { $0 == nft.id })
-    }
-
-    private func loadNfts() {
-        if !nftsId.isEmpty {
-            for id in nftsId where !nfts.contains(where: { $0.id == id }) {
-                loadNft(with: id)
-            }
-        }
-    }
-
-    private func loadNft(with id: String) {
-        view?.loadingDataStarted()
-        service.loadNft(id: id) { [weak self] result in
-            guard let self else { return }
-            switch result {
-            case .success(let nft):
-                DispatchQueue.main.async {
-                    self.view?.loadingDataFinished()
-                    self.nfts.append(nft)
-                    self.view?.reloadCollectionView()
-                }
-            case .failure:
-                let errorModel = ErrorModel(
-                    message: "Error.network"~,
-                    actionText: "Error.repeat"~
-                ) { [weak self] in
-                    guard let self else { return }
-                    self.loadNft(with: id)
-                }
-                DispatchQueue.main.async {
-                    self.view?.showError(errorModel)
-                }
-            }
-        }
-    }
-
-    func addOrDeleteFromCart(nft: Nft) {
-        if nftsInCart.contains(where: { $0.id == nft.id }) {
-            self.deleteFromCart(nft: nft)
-        } else {
-            addToCart(nft: nft)
-        }
-    }
 
     private func addToCart(nft: Nft) {
         if nftsInCart.isEmpty {
@@ -148,6 +78,79 @@ extension UsersCollectionPresenter: UsersCollectionPresenterProtocol {
                 assertionFailure("\(error)")
             }
         }
+    }
+
+    private func loadNfts() {
+        if !nftsId.isEmpty {
+            for id in nftsId where !nfts.contains(where: { $0.id == id }) {
+                loadNft(with: id)
+            }
+        }
+    }
+
+    private func loadNft(with id: String) {
+        view?.loadingDataStarted()
+        service.loadNft(id: id) { [weak self] result in
+            guard let self else { return }
+            switch result {
+            case .success(let nft):
+                DispatchQueue.main.async {
+                    self.view?.loadingDataFinished()
+                    self.nfts.append(nft)
+                    self.view?.reloadCollectionView()
+                }
+            case .failure:
+                let errorModel = ErrorModel(
+                    message: "Error.network"~,
+                    actionText: "Error.repeat"~
+                ) { [weak self] in
+                    guard let self else { return }
+                    self.loadNft(with: id)
+                }
+                DispatchQueue.main.async {
+                    self.view?.showError(errorModel)
+                }
+            }
+        }
+    }
+}
+
+// MARK: - UsersCollectionPresenterProtocol
+extension UsersCollectionPresenter: UsersCollectionPresenterProtocol {
+    func viewDidLoad() {
+        loadNfts()
+        updateCart()
+        loadProfile()
+    }
+
+    func setNftsId(nftsId: [String]) {
+        self.nftsId = nftsId
+    }
+
+    func nftsCount() -> Int {
+        return nfts.count
+    }
+
+    func nftForIndex(indexPath: IndexPath) -> Nft {
+        return nfts[indexPath.row]
+    }
+
+    // MARK: - cart
+    func cartContainsNft(nft: Nft) -> Bool {
+        nftsInCart.contains(where: { $0.id == nft.id })
+    }
+
+    func addOrDeleteFromCart(nft: Nft) {
+        if nftsInCart.contains(where: { $0.id == nft.id }) {
+            self.deleteFromCart(nft: nft)
+        } else {
+            addToCart(nft: nft)
+        }
+    }
+
+    // MARK: - update favourites
+    func isInFavourites(nft: Nft) -> Bool {
+        nftsInFavourites.contains(where: { $0 == nft.id })
     }
 
     func updateFavoriteNft(nftId: String) {
