+++ FakeNFT/Scenes /Statistics/UsersCollection/UsersCollectionPresenterProtocol.swift
@@ -1,18 +1,11 @@
 //
-//  UsersCollectionContract.swift
+//  UsersCollectionPresenterProtocol.swift
 //  FakeNFT
 //
-//  Created by Bogdan Fartdinov on 21.04.2024.
+//  Created by Bogdan Fartdinov on 25.04.2024.
 //
 
 import UIKit
-
-protocol UsersCollectionViewProtocol: AnyObject, ErrorView, LoadingView {
-    var presenter: UsersCollectionPresenterProtocol { get set }
-    func reloadCollectionView()
-    func loadingDataStarted()
-    func loadingDataFinished()
-}
 
 protocol UsersCollectionPresenterProtocol {
     var view: UsersCollectionViewProtocol? { get set }
